session-5/pyproject.toml
[project]
name = "session-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.4.10",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
